Remove unnecessary trait bound `From<u64>`
