+++ primitives/src/circuit/merkle_tree/mod.rs
@@ -447,7 +447,8 @@ mod test {
             constrain_sibling_order, Merkle3AryMembershipProofVar, MerkleTreeGadget,
         },
         merkle_tree::{
-            internal::MerkleNode, prelude::RescueMerkleTree, MerkleCommitment, MerkleTreeScheme,
+            internal::MerkleNode, prelude::RescueMerkleTree, AppendableMerkleTreeScheme as _,
+            MerkleCommitment, MerkleTreeScheme,
         },
         rescue::RescueParameter,
     };
