+++ primitives/src/merkle_tree/light_weight.rs
@@ -26,7 +26,7 @@ use num_traits::pow::pow;
 use serde::{Deserialize, Serialize};
 use typenum::Unsigned;
 
-impl_merkle_tree_scheme!(LightWeightMerkleTree, build_light_weight_tree_internal);
+impl_merkle_tree_scheme!(LightWeightMerkleTree);
 impl_forgetable_merkle_tree_scheme!(LightWeightMerkleTree);
 
 impl<E, H, I, Arity, T> AppendableMerkleTreeScheme for LightWeightMerkleTree<E, H, I, Arity, T>
@@ -37,6 +37,20 @@ where
     Arity: Unsigned,
     T: NodeValue,
 {
+    fn from_elems(
+        height: usize,
+        elems: impl IntoIterator<Item = impl Borrow<Self::Element>>,
+    ) -> Result<Self, PrimitivesError> {
+        let (root, num_leaves) =
+            build_light_weight_tree_internal::<E, H, I, Arity, T>(height, elems)?;
+        Ok(Self {
+            root,
+            height,
+            num_leaves,
+            _phantom: PhantomData,
+        })
+    }
+
     fn push(&mut self, elem: impl Borrow<Self::Element>) -> Result<(), PrimitivesError> {
         <Self as AppendableMerkleTreeScheme>::extend(self, [elem])
     }
