+++ primitives/tests/merkle_tree_hasher.rs
@@ -1,6 +1,8 @@
 use jf_primitives::{
     errors::PrimitivesError,
-    merkle_tree::{hasher::HasherMerkleTree, MerkleCommitment, MerkleTreeScheme},
+    merkle_tree::{
+        hasher::HasherMerkleTree, AppendableMerkleTreeScheme, MerkleCommitment, MerkleTreeScheme,
+    },
 };
 use sha2::Sha256;
